feat: Generate glue and replace custom module with it
Generated glue code and used the generated glue code to achieve the same
functionality as the handwritten `roc.rs` file for interfacing with the
roc platform. Generated code for the x86_64 target which is the target
this project is built for was editted due to limitations with the glue
generating code i.e. cannot generate code containing `Box`ed types